+++ packages/bridge-controller/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Export `calcNormalizedTokenAmount` and `calcAtomicTokenAmount` conversion utils
+- Export `AmountsAndAsset` type
+- Add utils to support fee validation when a quote's fees are denominated in multiple assets
+  - `hasSufficientGasForQuote` returns true if the wallet's balances are greater than or equal to the quote's network fees
 
 ### Changed